[Task FrameTransformation] Removing defining error

+++ app/FrameTransformation.cpp
@@ -45,8 +45,8 @@ std::vector<double> FrameTransformation::transformFrame(
     const std::vector<double> &position) {
     double mult_sum = 0;
     std::vector<double> finalPosition;
-    for (size_t i = 0; i < 4; i++) {
-      for (size_t j = 0; j < 4; j++) {
+    for (int i = 0; i < 4; i++) {
+      for (int j = 0; j < 4; j++) {
         mult_sum += transformationMatrix[i][j]*position[j];
       }
       finalPosition.push_back(mult_sum);
